+++ changelog.md
@@ -12,6 +12,7 @@
 -   `MapParse` Mirror names
 -   `WhatIf` command not properly quitting if a null value is given
 -   `Simulate` DT/NC/HT/DC no longer multiplies `overrideSpeed`
+-   `Simulate` custom speed from previous scores is ignored
 
 ### Changed